+++ app/components/settings/categories/WalletSettings.scss
@@ -1,4 +1,4 @@
-@import '../../themes/mixins/error-message';
+@import '../../../themes/mixins/error-message';
 
 .component {
   background-color: var(--theme-main-body-background-color);
